Added security app
- Added auth endpoints
- Removed some user endpoints
