+++ security/views/pair_token_view.py
@@ -1,13 +1,13 @@
 from drf_spectacular.utils import extend_schema
-from rest_framework_simplejwt.views import TokenRefreshView
+from rest_framework_simplejwt.views import TokenObtainPairView
 
-SCHEMA_NAME = "users"
+SCHEMA_NAME = "auth"
 
 
 @extend_schema(
     tags=[SCHEMA_NAME],
 )
-class RefreshTokenView(TokenRefreshView):
+class PairTokenView(TokenObtainPairView):
     # @extend_schema()
     def post(self, request, *args, **kwargs):
         return super().post(request, *args, **kwargs)
