+++ security/views/sign_up_view.py
@@ -7,7 +7,7 @@ from rest_framework.views import APIView
 
 from ..serializers import SignUpSerializer
 
-SCHEMA_NAME = "users"
+SCHEMA_NAME = "auth"
 
 
 @extend_schema(tags=[SCHEMA_NAME])
@@ -22,8 +22,7 @@ class SignUpView(APIView):
     )
     def post(self, request, *args, **kwargs):
         try:
-            data = request.data
-            serializer = self.serializer_class(data=data)
+            serializer = self.serializer_class(data=request.data)
             if not serializer.is_valid():
                 response = serializer.errors
                 return Response(response, status.HTTP_400_BAD_REQUEST)
@@ -31,8 +30,7 @@ class SignUpView(APIView):
             headers = self.get_success_headers()
             response = serializer.data
             return Response(response, status.HTTP_201_CREATED, headers=headers)
-        except Exception as e:
-            print(e)
+        except Exception:
             response = {
                 "title": "Internal error",
                 "message": "There was an error trying to sign you up.",
