+++ user/serializers/__init__.py
@@ -1,2 +1,1 @@
 from .profile_serializer import ProfileSerializer
-from .sign_up_serializer import SignUpSerializer
