+++ user/views/__init__.py
@@ -1,4 +1,1 @@
 from .profile_view import ProfileView
-from .sign_up_view import SignUpView
-from .pair_token_view import PairTokenView
-from .refresh_token_view import RefreshTokenView
